requirements.txt
selenium==4.4.3
webdriver-manager
pytest==7.1.3
allure-pytest==2.11.1
pytest-rerunfailures~=10.2
pytest-html-reporter~=0.2.9
pytest-check